Module adapton_lab::catalog::eg_oopsla2015

This 'list_mapimplementation and itsList<_>` datatype follow the code listing from "Incremental Computation with Names" (OOPSLA 2015).

Aside: Compared to the version of List<_> in the Adapton collections library, this version is simpler: It assumes only two constructors, Nil and Cons, and that every list element has an associated name and reference cell.

This generate function creates the three-element, three-name, three-ref-cell list in the first parts of Figures 1 and Figure 2.

This edit function implements the single insertion of Cons(2, ...) into the generated list, as per the second parts of Figures 1 and 2. (It only performs this one edit, and performs no other actions later). This code shows two ways of inserting a new element (and name and ref cell) into the list: The "functional" way, and the imperative way. The use of names here bridges the gap, permitting the functional approach to express the mutation in the imperative approach.

Structs

Archivist

Editor

Enums

List
Cons cells carry an element, name and reference cell for the rest of the list.

Functions

list_map
List map, as shown in 'Incremental Computation with Names', Section 2:

list_map_cons
This is the code that we memoize each time we see a name in a Cons cell. We identify this memo point using the names from the list.