Function adapton_lab::catalog::all_labs

pub fn all_labs() -> Vec<Box<Lab>>

all_labs: This is the master list of all labs in the current Adapton Lab catalog (which consists of this module, but which wraps the Adapton crate's collections module).

To add a new lab, just add a labdef! to the vec! in this definition. Doing so generally requires doing at least the following:

Add a new (empty struct) type that implements the Compute or ComputeDemand trait for some input- and output-type pair.

The input type must be a type parameter to some implementation of the Generate and Edit traits, which you must also provide to labdef!. Together, these trait implementations give the distribution of the input (how it is chosen, and how it changes, respectively).

The other arguments consist of the type of the Editor state (e.g., a counter of type usize in many cases), and a name and url to display in generated output. The URL should link to the rustdoc for this module, which in turn provides other related documentation about Adapton Lab and Adapton.