Module adapton_lab::catalog::hammer_s17_hw0

Hammer - CSCI 7000, Spring 2017

First homework assignment: #HW0

There are three functions below whose bodies consist of panic!("TODO"). Using the types listed in their declarations, implement these functions. You will find the list_map example from the oopsla2015_sec2 module helpful, as a guide.

Test the behavior of your filter function, for instance:

cargo run -- -L hammer-s17-hw0-filter

Structs

Editor
The Editor in this example generates a three-element initial list, then inserts an additional element. It's the same as oopsla2015_sec2::Editor.

RunFilter

RunReverse

RunSplit

Enums

List
Cons cells carry an element, name and reference cell for the rest of the list.

Functions

list_filter
List filter:

list_reverse
List reverse:

list_split
List split: