Struct adapton_lab::catalog::EditorOopsla2015Sec2

pub struct EditorOopsla2015Sec2 {}

Example from Incremental Computation with Names (2015), Section 2 (Figs 1 and 2)

This generate function creates the three-element, three-name, three-ref-cell list in the first parts of Figures 1 and Figure 2.

This edit function implements the single insertion of Cons(2, ...) into the generated list, as per the second parts of Figures 1 and 2. (It only performs this one edit, and performs no other actions later). This code shows two ways of inserting a new element (and name and ref cell) into the list: The "functional" way, and the imperative way. The use of names here bridges the gap, permitting the functional approach to express the mutation in the imperative approach.