Struct adapton_lab::catalog::ExampleCleanDirty

pub struct ExampleCleanDirty {}

Simple example to explain Adapton's dirtying + cleaning algorithms.

This example demonstrates how, when input changes, dirtying proceeds incrementally through the edges of the DCG, during cleaning. We show a particular input change for this DCG where a subcomputation h is never dirtied nor cleaned by change propagation. We show another change to the same input where this subcomputation h is eventually dirtied and cleaned by Adapton, though not immediately.

  ref cell
  inp
  [ 2 ]
    ^
    | force                                                         
    | 2                                                      
    |                 cell                                   cell
    |    alloc 4       b      force 4           alloc 4       c
   (g)-------------->[ 4 ]<--------------(h)--------------->[ 4 ]
    ^                                     ^
    | force                               | force h,
    | returns b                           | returns c
    |                                     |
   (f)------------------------------------+
    ^
    | force f,
    | returns cell c
    |
 (root of demand)

In this example DCG, thunk f allocates and forces two sub-computations, thunks g and h. The first consumes the input inp and produces an intermediate result (ref cell b); the second consumes this intermediate result and produces a final result (ref cell c), which both thunks h and f return as their final result.

When the input cell inp changes, e.g., from 2 to -2, thunks f and g are dirtied. Thunk g is dirty because it consumes the changed input. Thunk f is dirty because it demanded (consumed) the output of thunk g in the extent of its own computation. Importantly, thunk h is not immediately dirtied when inp changes.

In some very real sense, inp is an indirect ("transitive") input to thunk h. This fact may suggest that when inp is changed from 2 to -2, we should dirty thunk h immediately. However, thunk h is related to this input only by reading a different ref cell (ref cell b) that dependents, indirectly, on inp, via the behavior of thunk g, on which thunk h does not directly depend (e.g., thunk h does not force thunk g).

Rather, when thunk f is re-demanded (from the "root of demand", maybe a larger DCG or a user), it will necessarily perform a cleaning process (aka, "change propagation"), re-executing g, its immediate dependent, which is dirty. Since thunk g merely squares its input, and 2 and -2 both square to 4, the output of thunk g will not change in this case. Consequently, the consumers of cell b, which holds this output, will not be dirtied or re-executed. In this case, thunk h is this consumer. In situations like these, Adapton's dirtying + cleaning algorithms do not dirty nor clean thunk h. (For some other change, e.g., from 2 to 3, thunk h would eventually be dirtied and cleaned).

In sum, under this change, after f is re-demanded, the cleaning process will first re-execute g, the immediate consumer of inp. Thunk g will again allocate cell b to hold 4, the same value as before. It also yields this same cell pointer (to cell b). Consequently, thunk f is not re-executed, and is cleaned. Meanwhile, the outgoing (dependency) edges thunk of h are never dirited.