Struct adapton_lab::labdef::LabDef

pub struct LabDef<Input, EditSt, Output, Editor: Generate<Input> + Edit<Input, EditSt>, Archivist: ComputeDemand<Input, Output>> {
    pub identity: Name,
    pub url: Option<String>,
    pub editor: PhantomData<Editor>,
    pub archivist: PhantomData<Archivist>,
    pub input: PhantomData<Input>,
    pub editst: PhantomData<EditSt>,
    pub output: PhantomData<Output>,
}

lab definition: generic notion of an incremental computation that can be evaluated and tested. We instantiate this structure once for each test in our test suite. We implement the LabDef trait generically for this structure. See catalog module for example instances.

Fields
identity: Name url: Option<String> editor: PhantomData<Editor> archivist: PhantomData<Archivist> input: PhantomData<Input> editst: PhantomData<EditSt> output: PhantomData<Output>