Struct adapton_lab::labdef::TestComputer

pub struct TestComputer<Input, EditSt, Output, InputDist: Generate<Input> + Edit<Input, EditSt>, Computer: Compute<Input, Output>> {
    pub identity: Name,
    pub computer: PhantomData<Computer>,
    pub input: PhantomData<Input>,
    pub editst: PhantomData<EditSt>,
    pub inputdist: PhantomData<InputDist>,
    pub output: PhantomData<Output>,
}

Generic notion of an Incremental Computation to Evaluate and Test. We instantiate this structure once for each test in our test suite. We implement the LabDef trait generically for this structure.

Fields
identity: Name computer: PhantomData<Computer> input: PhantomData<Input> editst: PhantomData<EditSt> inputdist: PhantomData<InputDist> output: PhantomData<Output>