Trait adapton_lab::labdef::Compute

pub trait Compute<Input, Output> {
    fn compute(_: Input) -> Output;
}

Generic notion of a computation to run naively and incrementally. It has specific Input and Output types, and a way to compute the Output from the Input. See README.md for more.

Required Methods

fn compute(_: Input) -> Output