Module adapton_lab::labviz

This module of Adapton Lab extends Adapton's existing reflection (see adapton::engine::reflect) with further reflections. In particular, we produce HTML output structure, for human user interaction and consumption.

Structs

Div
The Div struct represents a restricted form of a <div> element in HTML. The field tag is a string, which corresponds to a a distinguished tag CSS class indicates the Rust datatype reflected into this Div. The other CSS classes hold bits that signal various subcases (e.g., of enums in the reflect module). Finally, for Rust structures that have subfields and/or substructure, the Div's extent field lists their reflections into Divs. In principle, the produced Div structure has an equivalent amount of information to the corresponding Rust datatype, and could be "parsed" back into this Rust datatype later (let's not do that, though!).

Traits

WriteHTML

Functions

div_of_edge

div_of_loc

div_of_name

div_of_oploc

div_of_path

div_of_succ

div_of_trace

style_string

write_test_results