Struct adapton_lab::labviz::Div

pub struct Div {
    pub tag: String,
    pub classes: Vec<String>,
    pub extent: Box<Vec<Div>>,
    pub text: Option<String>,
}

The Div struct represents a restricted form of a <div> element in HTML. The field tag is a string, which corresponds to a distinguished tag CSS class that indicates the Rust datatype reflected into this Div. The other CSS classes hold bits that signal various subcases (e.g., of enums in the reflect module). For Rust structures that have subfields and/or substructure, the Div's extent field lists their reflections into Divs. In principle, the produced Div structure has an equivalent amount of information to the corresponding Rust datatype, and could be "parsed" back into this Rust datatype later (let's not do that, though!). The text field is useful for Div versions of Names, for giving the text of the name.

Fields
tag: String classes: Vec<String> extent: Box<Vec<Div>> text: Option<String>