Struct adapton_lab::SampleParams

pub struct SampleParams {
    pub input_seeds: Vec<usize>,
    pub generate_params: GenerateParams,
    pub validate_output: bool,
    pub change_batch_size: usize,
}

Parameters for collecting a single sample. In addition to these parameters, the experiment maintains a Rng based on the input_seeds, below; this Rng is given to Edit::edit to generate psuedo-random edits, in batches. For each engine, this Rng is sequenced across successive samples. Given an input_seeds vector, there is one unique Rng sequence for each engine's sequence of samples.

Fields
input_seeds: Vec<usize>

We convert this seed into a random-number-generator before generating and editing.

generate_params: GenerateParams

Other parameters for generating the input.

validate_output: bool

Whether to validate the output after each computation using the naive and DCG engines

change_batch_size: usize

Size of each batch of changes.